objetivo: prioridade vira float, remove swap automático
- Remove lógica de swap em _aplicar_prioridade — prioridades duplicadas agora são permitidas
- Migra coluna prioridade de INTEGER para REAL no modelo
- Atualiza parse de int() para float() nas rotas add e edit
- Inputs HTML recebem step=0.1 para aceitar decimais

+++ app/models/geld_models.py
@@ -111,7 +111,7 @@ class Objetivo(Base):
     valor_inicial = Column(Numeric(15,2), nullable = False)
     data_inicial = Column(DateTime, nullable = False)
     data_final = Column(DateTime, nullable = False)
-    prioridade = Column(Integer, nullable=True)
+    prioridade = Column(Float, nullable=True)
 
 
     @property
